+++ src/detection/patterns/PatternPayload.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "PatternTypes.h"
-#include "PatternCandidate.h"
-#include "PatternResult.h"
-